requirements.txt
alembic==1.7.5
aniso8601==9.0.1
anyio==3.6.1
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
blinker==1.4
certifi==2022.5.18.1
click==8.0.3
coverage==6.1.2
dadata==21.10.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.9.0
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.2
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
pyparsing==2.4.7
pytest==6.2.5
pytest-httpx==0.21.0
pytz==2021.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
speaklater==1.3
SQLAlchemy==1.4.27
toml==0.10.2
tomli==2.0.1
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
